requirements.txt
openai
fastapi
uvicorn[standard]
chromadb
pdfplumber
python-dotenv
pydantic
streamlit